@sbarzowski 2019-02-28
Example benchmark:
```
std.foldl(function(e, res) e + res, std.makeArray(20000, function(i) "aaaaa"), "")
```

This version: 0.73s user 0.13s system 131% cpu 0.652 total
Current master: 1.64s user 1.89s system 106% cpu 3.306 total
C++ version: 1.24s user 0.61s system 98% cpu 1.872 total