@sparkprime 2019-05-08
This is a good idea.  We should tune it on some realistic benchmarks if possible.

I think having a native deep join that will do 2 passes over the tree will still be faster in the case of manifest functions, but obviously that is more of an expert approach.