@sbarzowski 2019-09-17
I rebased it and changed the names of string representations to something a little bit more concrete. 

@sparkprime Can we merge this?